Collection API parity, task 1 closed: hashing parity — upstream's `SIMD[_, _]` hasher leaf, keyed `AHasher[key: U256]`, container `Hashable`, the bytes overloads, and compile-time `hash`
The hasher protocol matched upstream's values but not its spellings: the leaf
was Mojito's own `_update_with_simd(mut self, value: UInt64)` (accepting user
hashers Mojo rejects), `AHasher` was key-less so `repr(set)` printed a
different text, the bytes overloads were missing, and only String/Optional/
Variant/Tuple were Hashable. Every item is now upstream's; each value was
checked against the audited build on both backends.

Protocol and leaf
- `_update_with_simd(mut self, value: SIMD[_, _])`: the elaborator desugars
  the wildcard parameter to an infer-only type parameter bounded by the
  hidden `$SIMD` and mints one clone per hashed vector type through the
  method-clone machinery (`_update_with_simd$y3:Int`) — the closed width-1
  set eagerly for every `Hasher` conformer (hashing reaches the hasher through
  erased paths, and the VM-CTFE subprogram has no discovery loop), wider
  vectors through a new checker demand channel (`hash_leaf_types`) the driver
  feeds back. The template body is a trap stub; every backend leaf dispatch
  computes the exact clone name (`simd_update_clone_name`), passes the value
  itself with `-0.0` folded (upstream's `SIMD.__hash__`), and a `Bool` leaf as
  `Scalar[DType.bool]`; `hash_bits` is gone. The retired `UInt64` spelling is
  a conformance error with a migration hint.
- `SIMD.to_bits[dtype]()` (new `MirInstr::SimdBitcast`, `simd.bits`) and the
  `.length` lane-count fold are intrinsics; width>1 vectors hash lane pairs.
  Natively a width-1 vector register is a scalar: indexing and reductions
  read it directly. The bundled bodies read lanes with `to_bits`/`.length`
  in `while` loops (`range` is not visible inside `std.hashlib`); pinned Mojo
  rejects a conformer that names the parameters, so they spell the wildcard.

Containers
- `List`, `Array`, `Set`, and `Dict` conform to `Hashable` conditionally on
  their elements with upstream's bodies (List/Array feed each element with no
  length prefix; Set XORs the default hasher's element hashes; Dict mixes
  per-entry hashes under a fresh caller-typed `H2`); `Variant` tags with a
  `UInt8`. A `where conforms_to(Self.T, Hashable)` assumption now proves
  `element.__hash__(hasher)` on an unbounded `T`, and the VM reifies a
  method-level `Hasher` binder inferred from its `mut` argument (`H2()` reads
  the hasher's struct name through the parameter's reference handle).

Keyed AHasher
- `struct AHasher[key: U256](Defaultable, Hasher)` with `comptime U256 =
  SIMD[DType.uint64, 4]`, both initializers mixing `Self.key` into the pi
  constants, `comptime default_hasher = AHasher[SIMD[DType.uint64, 4](0)]`,
  and `hash_seeded`/`hash_seeded_bytes` over the zero-keyed specialization.
  The compiler gained a compile-time SIMD value (`CtValue::Simd`: wrapped
  lanes, materialization, upstream's parameter rendering, MIR text round
  trip, VM conversions), SIMD-typed struct value parameters (a vector alias
  used as a bound folds to the parser's value-type spelling; the struct
  monomorphizes per application; `Self.key` bakes in clone bodies;
  `rewrite_type` folds an alias name to its bound type), compile-time SIMD
  construction and lane-wise `^ & | + - *` in the elaborator and the checker's
  constant evaluator, the single-argument struct application in an alias (an
  `Index` parse shape), and one canonical identity for a concrete application
  of a vector-keyed template — the mangled clone — shared by the checker's
  default fill, `ConstructTypeParam`, monomorphization, and the specialization
  oracle (which answers for a clone through `demangle_specialization`).
  `_unqualified_type_name` spells a minted clone with its baked values, so
  `repr(set)` prints `Hasher=AHasher[[0, 0, 0, 0] : SIMD[DType.uint64, 4]]`.

Bytes overloads
- `hash[HasherType: Hasher = default_hasher](bytes: ImmPointer[UInt8, _], n:
  Int)` and `hash_seeded_bytes(data, n, seed)`. A placeholder-origin pointer
  parameter of a free function resolves to the unsafe-any provenance at the
  alias's permission (any pointer binds it through `PointerOrigin::coerces_to`
  in both coercion predicates; the callee holds no loan); the bare
  `Pointer[T, _]` asks for the permission upstream infers. Span's raw
  constructor takes an immutable untracked pointer (a mutable one binds by
  capability drop), and a pointer alias annotation's overload symbol spells
  `UnsafePointer$` like every other pointer.

Compile-time hashing
- `comptime CT = hash[default_comp_time_hasher](Int(1))` (and the AHasher
  default) evaluates through VM CTFE: a type-parameterized free-function call
  with a constructible parameter routes through a synthesized checked entry
  (folded type arguments, typed literal arguments), the elaborator constructs
  typed scalars (`Int(1)`), the purity walk admits the hasher protocol's
  method calls and the bundled hashers' mixing steps, `to_bits`/`cast`,
  scalar-alias/`SIMD`/vector-alias constructions, and a vector-keyed
  template's constructors, and the VM-CTFE subprogram carries the keyed
  `default_hasher` clone at its template's position with its callee closure
  seeded from the template. Scalar, Bool, Float64, and string arguments fold;
  compile-time Dict/Set values are the new roadmap task 8.

Native lane
- With two `hash` overloads the elaborator no longer mints per-call `hash`
  clones, which exposed that the monomorphizer bound a supplied constructible
  type argument from the declaration default: positional compile-time
  arguments now skip inferred-only declarations, a reified type argument's
  string constant binds the type slot, and a spelling that names an
  enclosing binder (`hash[Self.H](key)` in Dict's erased bodies) resolves
  through the instance being specialized — which also closes the native
  residue where `Dict[K, V, SumHasher]` entries hashed under `AHasher`.

Fixtures and docs
- New `assets/ok/{hasher_upstream_spellings, container_hashable,
  hashlib_keyed_ahasher, hashlib_bytes_hash, comptime_hash,
  dict_hasher_forwarding}.mojo`, `conformance/fixtures/set_repr_keyed_hasher.mojo`,
  `assets/type_error/{hasher_uint64_leaf, list_element_not_hashable,
  pointer_placeholder_needs_permission}.mojo`; `hashlib_values` gains
  multi-lane/Float32/`to_bits` lines and `hasher_user_conformer` is
  respelled; hand-written manifest rows, exe ratchet 398 → 404, `cases.tsv`
  rows, and `run_compiled` golden pins in `tests/self_host_test.rs`.
- `docs/roadmap.md` task 1 deleted (tasks renumbered 1–7, task 8
  "Compile-time collections" added, the native `hash[Self.H]` residue
  removed), `docs/features.md`, `conformance/parity.tsv`,
  `docs/notes/hashlib-subset.md` (rewritten around the surviving
  `_folded_multiply` narrowing), `docs/symbol-map.md`, `docs/architecture.md`,
  `stdlib/README.md`, `CHANGELOG.md`.

In-session gate: `cargo build` on both lanes, `mojito run` on the touched
fixtures (VM and `--backend pliron`), the five golden tests, `cargo fmt`,
`git diff --check`. Suites, corpus, manifest regeneration, and Clippy are the
nightly gate's.

+++ stdlib/std/utils/variant.mojo
@@ -56,7 +56,7 @@ struct Variant[*Ts: AnyType](
         comptime for i in range(Self.Ts.length):
             comptime T = Self.Ts[i]
             if self._storage.isa[T]():
-                hasher.update(UInt64(i))
+                hasher.update(UInt8(i))
                 hasher.update(self._storage.unsafe_get[T]())
                 return
